// src/composables/games/tbh/enums/ending.ts
export enum TBH_ENDINGS {
  BAD_END_00 = "BAD_END_00",
  BAD_END_01 = "BAD_END_01",
  BAD_END_02 = "BAD_END_02",
  BAD_END_03 = "BAD_END_03",
  BAD_END_04 = "BAD_END_04",
  HAPPY_END_01 = "HAPPY_END_01",
  HAPPY_END_02 = "HAPPY_END_02",
  TRUE_END_01 = "TRUE_END_01",
  TRUE_END_02 = "TRUE_END_02",
  TRUE_END_03 = "TRUE_END_03",
  TRUE_END_04 = "TRUE_END_04",
  TRUE_END_05 = "TRUE_END_05",
}